Rebinning work
Based off of other projects with CNN, since the shape of the image matters a lot, this rebinning greatly reduces the distortion in the image that is used. Current problem is a line of alternating 22 pixels down the center of the camera that have nothing in them and do not correspond to pixels in the camera. But that should not matter as much as the shape of events should be much better preserved now.