More Rebinning, and use the current one for preprocessing
New one should be better, once I figure out how to tellesate teh hexagons into the detector shape, assign the CHID to it, and then create the square overlay on top of it. Thenartial hexagons to each square, keeping the information of the detector without the wholes from teh current one. Issue would be needing to change the preprocessor for everything, but should be doable. Would then be able to run both gamma/hadron separation and disp method better, I think. Imrpvoements in that should allow for better Li and Ma Significance for the final image.