Fix the flipping in the output
The smaller the binning, the more accurate output, obviously, so might try with binning of 4 or less, 8 or 9 are too fuzzy to work as well as hoped. 3 and 2 give great detail but are multiple times larger than the other ones. Also seems to be some type of fringing going on, almost like chromatic oberattion around the edges, but in all cases, better than the previous time