Various tweaks to sourceFinding Gamma/Hadron Sep
Acc on training is up to 86 percent, but on validation is still around 67 percent, so not great.

Hyperparameters seem to have a large effect, so gonna try to automate changing them and making networks based off the other two papers, not just Jan's work, since the other netowrks seem to have worked better.