Update preprocessors
Now should be able to do the job of the individual ones

Removed original ones now since these should work and are more consistent with the other parts of factnn

Dynamically generate the rebinning now, instead of from file. Doing this because the rebinning generation takes relatively little time compared to the preprocessing anyway, and not need to deal with another file.