Fix all preprocessors failing to build HDF5 files
Now they work with the FACT tools v1.0.0 dl2 files and successfully create the datasets

Also changed how the image matricies are created, removed the - 30 that was there from previous code for no real good reason. Instead now there is a check to make sure that the last index is less than the length of the data cube. If it is not, nothing happens, if it is, its value is added dependent on the mapping as before. This is to allow for smaller HDF5 datasets that do not need to store the whole 100 time slice chunk, instead just can pare it down to the first 50 or 25 or whatever is required, creating massive memory savings.