Fix errors in flow for energy and source files
On SSD, once it starts, the flow from directory does not seem to be as slow as I would have thought and the extra freedom from change all parameters easily outweighs the majority of the problems. Only real issue now is that events in individual files are clumped and not complete randomness between those types of events then. So still some benefit to HDF5, but not as much as I would have thought