Start on combining time dimension
Now can clunkily sum up parts of the time dimension to make the files load quicker. In addition, the time dimension can also be put as the channels for streaming in files now, so that there is an easier possibility of using other neural network architectures that only take 4D tensors instead of 5D ones.

API is still work in progress.