Misc updates and tests
Something seems to be wrong in the streaming framework, since it gets nearly perfect results every time on the simulated data, even when down to two neurons. While it does work correctly when the data is collapsed from the first 30 frames, where nearly all images have zero photons recorded, it does make it suspicous as to what should be done about it and what I am doing wrong with it.