Fix generator for None proton preprocessor
Still need to do more error checks and allow for setting the final_slices in the config instead of manually as of now.

Various changes to the examples codes as I still try to find out what's going wrong with the code. Validation error still much lower than training error for the Separation codes, although now able to get the worst possible loss when the network is down to 1 neuron per layer. With 2 to 4 per layer the loss drops to about 0.2, but validation loss is down an order of magnitude, so not sure what is happening there. Testing out the energy flow from files now to see if same oddly perfect results occur in a different type of example.