Start work on Sequence generator
Possible benefits and drawbacks listed in the sequence generator. Thinking need to split up the generators based on use case, so base_generator is used for the infintely looping data, single threaded as well works well, while Sequence_Generator will be for the multiprocessing, with each call getting elements from a single file, either with sampling for training, or just the first batch size elements for validation. Prediction would need a different generator. Finally, one that takes files with all of the files consisting of a single image and the necessary metadata, like in most examples of flowing from directories. This is used for getting the benefits of streaming in files while being able to know how many events are in each dataset, better augmentation and shuffling, etc. Although slower and probably going to take up more space.