Start on cleaning images
First step is adding support for using the DBSCAN clustering of the point clouds. Prelim idea works now, just need to add it more to the rest of the preprocessors and reprocess the eventfiles with it. Other cleaning or selection criteria might be added in the future.

Idea is to see if after using the clustering whether the images result in better classification, training, or regression using the different current models.

Also started slightly in adding Gaussian kernels to the rebinning, based on Holch's 2016 Msc thesis, but not sure whether that should help a lot or not as in the later 2017 paper using rebinning, oversampling, Gaussian, and others he found that Gaussian and oversampling results in the most distortion for an event compared to rebinning.

Attempts to redefine binning so that the size is the size in pixels, i.e. rebin_size=100 would be a 100x100 grid, instead of the square size, is seeming like it would be more useful and potentially more intuitive.

All these transformations and cleaning still need to have sanity checks applied to them to make sure they still match the original images correctly.